:sparkles: Move queue assign controls out of the assignee hover panel
