+++ components/assignments/QueueAssigneeStatus.tsx
@@ -1,20 +1,14 @@
 'use client'
 
 import { useState } from 'react'
-import { PlusIcon } from '@heroicons/react/24/outline'
 import { ToolsOzoneQueueDefs, ToolsOzoneTeamDefs } from '@atproto/api'
 import { formatDistanceToNowStrict } from 'date-fns'
 import { toast } from 'react-toastify'
-import {
-  useLabelerAgent,
-  usePermission,
-  useServerConfig,
-} from '@/shell/ConfigurationContext'
+import { useServerConfig } from '@/shell/ConfigurationContext'
 import { displayError } from '@/common/Loader'
 import { ConfirmationModal } from '@/common/modals/confirmation'
-import { useAssignQueue, useUnassignQueue } from './useAssignments'
+import { useUnassignQueue } from './useAssignments'
 import { Assignee } from './Assignee'
-import { AssigneeSearchPopover } from './AssigneeSearchPopover'
 
 interface QueueAssigneeStatusProps {
   queueId: number
@@ -25,12 +19,9 @@ export function QueueAssigneeStatus({
   queueId,
   assignments,
 }: QueueAssigneeStatusProps) {
-  const canManageQueues = usePermission('canManageQueues')
   const { role } = useServerConfig()
   const isAdmin = role === ToolsOzoneTeamDefs.ROLEADMIN
-  const myDid = useLabelerAgent().did
 
-  const { mutate: assignQueue } = useAssignQueue()
   const [pendingUnassign, setPendingUnassign] =
     useState<ToolsOzoneQueueDefs.AssignmentView | null>(null)
   const [unassignError, setUnassignError] = useState<string | undefined>()
@@ -53,14 +44,6 @@ export function QueueAssigneeStatus({
       setUnassignError(displayError(err))
     },
   })
-
-  const isAssignedToMe = assignments.some((a) => a.did === myDid)
-  const showAssignSelf =
-    !isAdmin && !isAssignedToMe && canManageQueues && myDid
-  const showEmptyIndicator =
-    assignments.length === 0 && !isAdmin && !showAssignSelf
-
-  const assignedDids = assignments.map((a) => a.did)
 
   const handleAssigneeClick = (a: ToolsOzoneQueueDefs.AssignmentView) => {
     setUnassignError(undefined)
@@ -89,25 +72,7 @@ export function QueueAssigneeStatus({
           onRemove={isAdmin ? () => handleAssigneeClick(a) : undefined}
         />
       ))}
-      {isAdmin && (
-        <AssigneeSearchPopover
-          excludeDids={assignedDids}
-          onSelect={(did) => {
-            assignQueue({ did, queueId })
-          }}
-        />
-      )}
-      {showAssignSelf && (
-        <button
-          type="button"
-          onClick={() => assignQueue({ did: myDid, queueId })}
-          className="text-xs text-indigo-600 dark:text-teal-400 hover:underline flex items-center gap-1"
-        >
-          <PlusIcon className="h-4 w-4" />
-          Assign to me
-        </button>
-      )}
-      {showEmptyIndicator && (
+      {assignments.length === 0 && (
         <p className="text-xs text-gray-400 dark:text-gray-500">-</p>
       )}
